Hide UserAgentParserBuilder behind builder method
