+++ nix/shell.nix
@@ -6,10 +6,15 @@ mkShell {
 
   nativeBuildInputs = [
     (rust-bin.stable.latest.default.override {
-      extensions = [ "rust-src" ];
+      extensions = [
+        "rust-src"
+        "rust-analysis"
+        "clippy"
+      ];
     })
     cargo-criterion
     cargo-edit
+    cargo-watch
     gnuplot
   ];
 }
